refactor(api): error file
