+++ api/src/worker/errors.ts
@@ -1,0 +1,6 @@
+export class WorkerRetryableError extends Error {
+  constructor(reason: string) {
+    super(reason);
+    this.name = "WorkerRetryableError";
+  }
+}
